feat(testing): Add free functions for simulating WSGI requests
Add free simulate_request, etc. functions that can be used without
requiring a unittest-style OO/inheritance model. Most notably, this
makes it easier to use pytest's more functional style when testing
falcon-based apps.

As part of this work, refactor the existing WSGI simulation logic
to layer on top of the free functions. Don't break backwards
compatibility.

Partially implements #16
