+++ doc/api/testing.rst
@@ -3,20 +3,11 @@
 Testing
 =======
 
-.. autoclass:: falcon.testing.TestCase
-    :members:
-
-.. autoclass:: falcon.testing.Result
-    :members:
-
-.. autoclass:: falcon.testing.SimpleTestResource
-    :members:
-
-.. autoclass:: falcon.testing.StartResponseMock
-    :members:
-
 .. automodule:: falcon.testing
-    :members: capture_responder_args, rand_string, create_environ
+    :members: Result, TestCase, SimpleTestResource, StartResponseMock,
+        simulate_request, simulate_get, simulate_head, simulate_post,
+        simulate_put, simulate_options, simulate_patch, simulate_delete,
+        capture_responder_args, rand_string, create_environ
 
 Deprecated
 ----------
