+++ falcon/testing/__init__.py
@@ -37,8 +37,9 @@ For additional examples, see also Falcon's own test suite.
 
 # Hoist classes and functions into the falcon.testing namespace
 from falcon.testing.base import TestBase  # NOQA
+from falcon.testing.client import *  # NOQA
 from falcon.testing.helpers import *  # NOQA
 from falcon.testing.resource import capture_responder_args  # NOQA
 from falcon.testing.resource import SimpleTestResource, TestResource  # NOQA
 from falcon.testing.srmock import StartResponseMock  # NOQA
-from falcon.testing.test_case import Result, TestCase  # NOQA
+from falcon.testing.test_case import TestCase  # NOQA
